+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: jbarbay <jbarbay@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/12 18:36:30 by jbarbay           #+#    #+#             */
-/*   Updated: 2024/03/11 14:45:16 by jbarbay          ###   ########.fr       */
+/*   Updated: 2024/03/11 16:53:11 by jbarbay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,10 @@
 # include <readline/readline.h>
 # include <readline/history.h>
 # include "../libft/libft.h"
+# include "structures.h"
 # include "parsing.h"
-# include "exec.h"
 # include "builtins.h"
+# include "exec.h"
 # include <stdio.h>
 # include <stdlib.h>
 # include <signal.h>
