+++ srcs/builtins/call_execve.c
@@ -50,6 +50,7 @@ void	ft_execve(char **cmd, t_data *data)
 	if (!env_value)
 	{
 		path = cmd[0];
+		data->exit_code = 1;
 		ft_putendl_fd("env: No such file or directory", STDOUT_FILENO);
 	}
 	else
@@ -57,12 +58,14 @@ void	ft_execve(char **cmd, t_data *data)
 	if (!path)
 	{
 		ft_putstr_fd(cmd[0], 2);
+		data->exit_code = 1;
 		ft_putendl_fd(": No such file or directory", 2);
 	}
 	if (execve(path, cmd, data->envp) == -1)
 	{
 		ft_putstr_fd(cmd[0], 2);
 		ft_putstr_fd(": command not found\n", 2);
+		data->exit_code = 127;
 		exit(EXIT_FAILURE);
 	}
 }
